+++ Framework/Configuration.php
@@ -101,6 +101,9 @@ class Configuration {
         if(getenv('DEBUG_SQL')) {
             self::$parameters['debug_sql'] = boolval(getenv('DEBUG_SQL'));
         }
+        if(getenv('DEBUG_INFLUXDB')) {
+            self::$parameters['debug_influxdb'] = boolval(getenv('DEBUG_INFLUXDB'));
+        }
         if(!isset(self::$parameters['smtp_from'])) {
             self::$parameters['smtp_from'] = 'donotreply@pfm.org';
         }
@@ -153,6 +156,16 @@ class Configuration {
         }
         if(getenv('PFM_AMQP_PASSWORD')) {
             self::$parameters['amqp_password'] = getenv('PFM_AMQP_PASSWORD');
+        }
+
+        if(getenv('PFM_INFLUXDB_URL')) {
+            self::$parameters['influxdb_url'] = getenv('PFM_INFLUXDB_URL');
+        }
+        if(getenv('PFM_INFLUXDB_TOKEN')) {
+            self::$parameters['influxdb_token'] = getenv('PFM_INFLUXDB_TOKEN');
+        }
+        if(getenv('PFM_INFLUXDB_ORG')) {
+            self::$parameters['influxdb_org'] = getenv('PFM_INFLUXDB_ORG');
         }