+++ CHANGELOG.md
@@ -6,11 +6,13 @@
 * New env variables:
   * DEBUG: activate logs with debug level
   * DEBUG_SQL: activate sql logs with debug level
+  * DEBUG_INFLUXDB: if using influxdb, set debug log level
   * PFM_MODE: [prod(default)|dev|test] in test, activate browser dev tab,
     in mode test disable html rendering and return template vars on controller call.
   * PFM_ADMIN_EMAIL: default super admin contact email
   * PFM_ADMIN_PASSWORD: default super admin user password (install only)
   * PFM_AMQP_HOST, PFM_AMQP_PORT, PFM_AMQP_USER, PFM_AMQP_PASSWORD: rabbitmq settings (optional)
+  * PFM_INFLUXDB_URL, PFM_INFLUXDB_TOKEN, PFM_INFLUXDB_ORG: influxdb settings
 
 * New config parameters:
-  * sql_debug: combined with DEBUG mode, trace sql queries
+  * same as env variables, without PFM_PREFIX and in lowercase